requirements.txt
anyio==4.4.0
ariadne==0.23.0
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
graphql-core==3.2.4
idna==3.9
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pymongo==4.8.0
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
Werkzeug==3.0.4
